Reexport some lang::ops instead of stdlib wrappers
